Improve the Blackboard LTI 1.3 upgrade management command
In the case that it's run after some users have already launched
upgraded LTI 1.3 links and so have new user accounts, it can merge them
with the original user accounts created by the LTI 1.1 consumer.

It takes the ID of the LTI 1.3 consumer as a required argument, to use
in the LTI_13_UserAlias objects.