keep a record of when scores are reported back to the consumer
Each time a student's score is reported to the consumer, a
UserScoreReported object is created. These are linked to a user, a
resource, an attempt and possibly a ReportProcess if all the resource's
scores were being reported back.
It records the awarded and maximum scores reported back, the completion
status, and any error message produced during the report.