Resource.part_hierarchy uses the exam definition
Previously, a Resource's part hierarchy was computed from the SCORM
elements of attempts at that resource. That meant that the hierarchy
can't be worked with until there is at least one attempt, and attempts
can't differ in their structure.

This replaces that routine with one that looks at the exam's definition.
It still doesn't cope with randomised order or different structures
resulting from things like explore or diagnostic mode, but at least it
can be determined without waiting for an attempt.

The form to discount parts is always available, fixing #344