Starting work on context summaries
A context summary page is a view of some of the resources in a context.

When students view it, they're shown for each resource their completion
status and score, and an optional total score for all the resources.

The total score can be either the sum of the scaled scores, the sum of
the raw scores, the number of resources completed or the number of
resources with 100% score.

The intention is to embed these in Canvas pages.
To do that, you have to insert a Numbas link from the Canvas editor,
then replace the link with an iframe whose source is the same URL as the
link's href. I haven't worked out how to persuade Canvas to just make
the iframe.

This commit also changes the deep link screen, to show existing
resources in a table instead of a list. I think it's a bit more compact.

see #292