Change to the logic around Attempt.get_end_time
If the student re-opens their attempt but doesn't change anything, we
think it's unfair to apply the full late work penalty.

This changes the logic to use the first time that the attempt was ended
after the last suspend data change. So if the student re-opens the
attempt and changes nothing, their original completion time will be
used.

If the student re-opens the attempt and leaves it open until the
"available until" time, it'll be automatically ended and the
`x.end_time` value will be set, so it'll look like they finished at the
"available until" time.