working on adding a minimum version check for lockdown apps
We want to set a minimum version of the lockdown app that students must
have, so we can deprecate old versions and ensure compatibility with the
Numbas runtime.

The immediate motivation is that the new Numbas default theme will use
CSS features that weren't supported by the old versions of the Numbas
lockdown app or SEB.

This also consolidates a lot of the lockdown app launch logic in a class
called numbas_lti.lockdown_app.LockdownApp, which has subclasses for the
Numbas app and SEB, as well as a trivial one for "no lockdown app
required".

TODO:

    * Check version numbers on the iOS and Android Numbas apps
    * Check how SEB puts its version number in the user agent