Improve the estimation of max_score
Resource.estimate_max_score only tried to look at the first attempt's max score. If that attempt was broken, then the whole resource would seem to not have a max score.

This changes the method to look through all the attempts and take the first one with a non-zero max score.