replace ScormElement signal handlers with functions run after save_scorm_data
I noticed a lot of foreign key constraint violations to do with the Attempt.completion_status_element field.

I think that the problem is that the SCORM elements are saved as part of an atomic transaction, and the scorm_set_completion_status task, triggered by a signal handler, could run before that transaction has finished.

Rather than increasing the delay on the task, I think the correct change is to not use signal handlers, and instead move the code that triggers those tasks to the end of scorm_set_completion_status, after the transaction is complete.