Re-marking: load extensions, and catch Numbas.schedule.halt
The re-mark page wasn't loading extensions, so any exams using extensions would just silently wait forever to load.
Now it loads extensions/extensions.json from the extracted package and inserts the extension scripts in the page.

Maybe we should just use the <numbas-exam> element, and insist that all themes support it, so the extension-loading code isn't duplicated.


This also overrides `Numbas.schedule.halt` at the top of `remark_session`, so if it's called, we can show the error to the user.

Ideally, Numbas.schedule.halt would emit a signal that we can listen for, instead.

Actually, instead of calling Numbas.schedule.halt, we should have a per-exam "quit everything" method, instead of a global function.