Only save SCORM data to local storage if the Commit method has been called
We had an annoyingly rare bug where students would launch the same attempt twice, before beginning the exam. If there was SCORM data saved to local storage but not saved to the server, then it would be restored on loading the attempt, but it would still be an 'ab-initio' entry.

This commit adds a check for this data in local storage when it's loaded, and changes the launch to 'resume' if so.

But that shouldn't happen any more, because it also only saves the data to localstorage once the SCORM API's `Commit` has been called once.