actually set the default timeout for requests!
I was using `kwargs.setdefault`, but the key `"timeout"` is set, to `None`!

This changes it to check `kwargs.get('timeout') is None`, so the timeout is set.

And I've checked it by using a session to make a request that I know will hang until the timeout:
    from numbas_lti.requests_session import get_session
    s = get_session()
    s.get('http://192.168.1.70:11000')

(192.168.1.70 is the IP of a device on my network which doesn't have port 11000 open)